Spark-ECS

+++ server/src/io/anuke/koru/server/ConnectionInfo.java
@@ -4,11 +4,12 @@ import com.esotericsoftware.kryonet.Connection;
 
 public class ConnectionInfo{
 	private static int lastid;
+	
 	public final int id;
-	public final long playerid;
+	public final int playerid;
 	public final Connection connection;
 	
-	public ConnectionInfo(long playerid, Connection connection){
+	public ConnectionInfo(int playerid, Connection connection){
 		this.id = lastid++;
 		this.playerid = playerid;
 		this.connection = connection;
